requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
scikit-learn==1.4.2
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9
